feat: port `extend_public_ids.py` script into `aea hash all` command
